+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [next]
 
 - fix: hide Tab's close button when `onClosed` is null
+- feat: Add `TextBox.cursorOpacityAnimates` (defaults to `FluentThemeData.cursorOpacityAnimates`, which defaults to `false`); default setting improves CPU/GPU efficiency while TextBox has focus ([#1164](https://github.com/bdlukaa/fluent_ui/issues/1164))
 
 ## 4.10.0